+++ Vault/src/utopia/vault/nosql/read/linked/CombiningDbRowReader.scala
@@ -2,7 +2,9 @@ package utopia.vault.nosql.read.linked
 
 import utopia.flow.collection.immutable.Empty
 import utopia.vault.model.immutable.{Row, Table}
+import utopia.vault.model.template.ConditionallyJoinable
 import utopia.vault.nosql.read.DbRowReader
+import utopia.vault.sql.Condition
 
 import scala.util.Try
 
@@ -14,21 +16,23 @@ object CombiningDbRowReader
 	  * @param left Reader used for parsing left side items
 	  * @param right Reader used for parsing joined items
 	  * @param bridges Joins to perform between the left and the right target (default = empty)
-	  * @param merge A function for merging both sides together
+	  * @param joinConditions Conditions that must be met in order for joining to occur (default = empty)
+	 * @param merge A function for merging both sides together
 	  * @tparam L Type of the parsed left-side items
 	  * @tparam R Type of the parsed right-side items
 	  * @tparam A Type of the merge results
 	  * @return A DB reader that combines the results of the specified two readers
 	  */
-	def apply[L, R, A](left: DbRowReader[L], right: DbRowReader[R], bridges: Seq[Table] = Empty)
+	def apply[L, R, A](left: DbRowReader[L], right: DbRowReader[R], bridges: Seq[Table] = Empty,
+	                   joinConditions: Seq[Condition] = Empty)
 	                  (merge: (L, R) => A): CombiningDbRowReader[L, R, A] =
-		new _CombiningDbRowReader[L, R, A](left, right, bridges, merge)
+		new _CombiningDbRowReader[L, R, A](left, right, bridges, joinConditions, merge)
 	
 	
 	// NESTED   ---------------------------
 	
 	private class _CombiningDbRowReader[L, R, A](left: DbRowReader[L], right: DbRowReader[R],
-	                                             bridges: Seq[Table], f: (L, R) => A)
+	                                             bridges: Seq[Table], joinConditions: Seq[Condition], f: (L, R) => A)
 		extends CombiningDbRowReader[L, R, A](left, right, bridges)
 	{
 		override protected def combine(left: L, right: R): A = f(left, right)
@@ -41,8 +45,9 @@ object CombiningDbRowReader
   * @since 10.07.2025, v1.22
   */
 abstract class CombiningDbRowReader[L, R, +A](left: DbRowReader[L], right: DbRowReader[R],
-                                              bridges: Seq[Table] = Empty)
-	extends JoiningDbReader[L, R, A](left, right, bridges) with DbRowReader[A]
+                                              bridges: Seq[Table] = Empty, joinConditions: Seq[Condition] = Empty)
+	extends JoiningDbReader[L, R, A](left, right, bridges, joinConditions = joinConditions) with DbRowReader[A]
+		with ConditionallyJoinable[DbRowReader[A]]
 {
 	// ABSTRACT -------------------------
 	
@@ -59,4 +64,7 @@ abstract class CombiningDbRowReader[L, R, +A](left: DbRowReader[L], right: DbRow
 	
 	override def shouldParse(row: Row): Boolean = left.shouldParse(row) && right.shouldParse(row)
 	override def apply(row: Row): Try[A] = left(row).flatMap { left => right(row).map { combine(left, _) } }
+	
+	override def onlyJoinIf(condition: Condition) =
+		CombiningDbRowReader(left, right, bridges, joinConditions)(combine)
 }