+++ Vault/src/utopia/vault/nosql/read/linked/MultiLinkedDbReader.scala
@@ -4,7 +4,7 @@ import utopia.flow.collection.immutable.Empty
 import utopia.vault.model.immutable.{Row, Table}
 import utopia.vault.nosql.read.DbRowReader
 import utopia.vault.nosql.read.parse.ParseGroupedRows
-import utopia.vault.sql.JoinType
+import utopia.vault.sql.{Condition, JoinType}
 import utopia.vault.sql.JoinType.Inner
 
 import scala.util.Try
@@ -18,6 +18,7 @@ object MultiLinkedDbReader
 	 * @param left The primary reader
 	 * @param right The secondary reader
 	 * @param bridges Joins to perform between the left and the right target (default = empty)
+	 * @param joinConditions Conditions that must be met in order for joining to occur (default = empty)
 	 * @param neverEmptyRight Whether to never expect a one-to-none connection.
 	 *                        If true, inner joining will be used.
 	 *                        Default = false = Left join will be used in order to accommodate the one-to-none case.
@@ -28,16 +29,18 @@ object MultiLinkedDbReader
 	 * @return A new database reader
 	 */
 	def apply[L, R, A](left: DbRowReader[L], right: DbRowReader[R],
-	                   bridges: Seq[Table] = Empty, neverEmptyRight: Boolean = false)
+	                   bridges: Seq[Table] = Empty, joinConditions: Seq[Condition] = Empty,
+	                   neverEmptyRight: Boolean = false)
 	                  (f: (L, Seq[R]) => A): MultiLinkedDbReader[L, R, A] =
-		new _MultiLinkedDbReader[L, R, A](left, right, bridges, f, neverEmptyRight)
+		new _MultiLinkedDbReader[L, R, A](left, right, bridges, joinConditions, f, neverEmptyRight)
 	
 	
 	// NESTED   ----------------------------
 	
 	private class _MultiLinkedDbReader[L, R, A](left: DbRowReader[L], right: DbRowReader[R],
-	                                            bridges: Seq[Table], f: (L, Seq[R]) => A, neverEmptyRight: Boolean)
-		extends MultiLinkedDbReader[L, R, A](left, right, bridges, neverEmptyRight)
+	                                            bridges: Seq[Table], joinConditions: Seq[Condition],
+	                                            f: (L, Seq[R]) => A, neverEmptyRight: Boolean)
+		extends MultiLinkedDbReader[L, R, A](left, right, bridges, joinConditions, neverEmptyRight)
 	{
 		override protected def combine(left: L, right: Seq[R]): A = f(left, right)
 	}
@@ -51,8 +54,10 @@ object MultiLinkedDbReader
  * @since 10.07.2025, v1.22
  */
 abstract class MultiLinkedDbReader[L, R, A](left: DbRowReader[L], right: DbRowReader[R],
-                                   bridges: Seq[Table] = Empty, neverEmptyRight: Boolean = false)
-	extends JoiningDbReader[L, R, A](left, right, bridges, if (neverEmptyRight) Inner else JoinType.Left)
+                                            bridges: Seq[Table] = Empty, joinConditions: Seq[Condition] = Empty,
+                                            neverEmptyRight: Boolean = false)
+	extends JoiningDbReader[L, R, A](left, right, bridges, if (neverEmptyRight) Inner else JoinType.Left,
+		joinConditions)
 		with ParseGroupedRows[A]
 {
 	// ABSTRACT ----------------------------